# apps/web/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "AvailabilityCommand": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "windows": {
            "items": {
              "$ref": "#/components/schemas/Interval"
            },
            "maxItems": 400,
            "title": "Windows",
            "type": "array"
          }
        },
        "required": [
          "expected_revision",
          "windows"
        ],
        "title": "AvailabilityCommand",
        "type": "object"
      },
      "AvailabilityResponse": {
        "additionalProperties": false,
        "properties": {
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "timezone": {
            "title": "Timezone",
            "type": "string"
          },
          "windows": {
            "items": {
              "$ref": "#/components/schemas/Interval"
            },
            "title": "Windows",
            "type": "array"
          }
        },
        "required": [
          "windows",
          "timezone",
          "revision"
        ],
        "title": "AvailabilityResponse",
        "type": "object"
      },
...
      "FixedEventCreate": {
        "additionalProperties": false,
        "properties": {
          "end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "$ref": "#/components/schemas/LocalTime"
              }
            ],
            "title": "End"
          },
          "expected_revision": {
            "minimum": 0,
            "title": "Expected Revision",
            "type": "integer"
          },
          "start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "$ref": "#/components/schemas/LocalTime"
              }
            ],
            "title": "Start"
          },
          "title": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "expected_revision",
          "start",
          "end",
          "title"
        ],
        "title": "FixedEventCreate",
        "type": "object"
      },
      "FixedEventPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FixedEventResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "revision"
        ],
        "title": "FixedEventPage",
        "type": "object"
      },
...
      "FixedEventResponse": {
        "additionalProperties": false,
        "properties": {
          "end": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "$ref": "#/components/schemas/LocalTime"
              }
            ],
            "title": "End"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "start": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "$ref": "#/components/schemas/LocalTime"
              }
            ],
            "title": "Start"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "start",
          "end",
          "id",
          "title",
          "revision"
        ],
        "title": "FixedEventResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RevisionCommand": {
        "additionalProperties": false,
        "properties": {
          "expected_revision": {
            "minimum": 0,
            "title": "Expected Revision",
            "type": "integer"
          }
        },
        "required": [
          "expected_revision"
        ],
        "title": "RevisionCommand",
        "type": "object"
      },
      "RevisionResponse": {
        "additionalProperties": false,
        "properties": {
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "revision"
        ],
        "title": "RevisionResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Adaptive Planner",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/availability": {
      "get": {
        "operationId": "get_availability_api_v1_availability_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Get Availability"
      },
      "put": {
        "operationId": "set_availability_api_v1_availability_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AvailabilityCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AvailabilityResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Availability"
      }
    },
...
    "/api/v1/fixed-events": {
      "get": {
        "operationId": "list_events_api_v1_fixed_events_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixedEventPage"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Events"
      },
      "post": {
        "operationId": "create_event_api_v1_fixed_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FixedEventCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixedEventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Event"
      }
    },
    "/api/v1/fixed-events/{event_id}": {
      "delete": {
        "operationId": "delete_event_api_v1_fixed_events__event_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Event"
      },
      "patch": {
        "operationId": "patch_event_api_v1_fixed_events__event_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "event_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Event Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FixedEventPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FixedEventResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Event"
      }
    },
...
    "/api/v1/tasks": {
      "get": {
        "operationId": "list_tasks_api_v1_tasks_get",
        "parameters": [
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Tasks"
      },
      "post": {
        "operationId": "create_task_api_v1_tasks_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Task"
      }
    },
    "/api/v1/tasks/{task_id}": {
      "get": {
        "operationId": "get_task_api_v1_tasks__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Task"
      },
      "patch": {
        "operationId": "patch_task_api_v1_tasks__task_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskPatch"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Patch Task"
      }
    },
    "/api/v1/tasks/{task_id}/cancel": {
      "post": {
        "operationId": "cancel_task_api_v1_tasks__task_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RevisionCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Task"
      }
    },
    "/api/v1/tasks/{task_id}/dependencies": {
      "delete": {
        "operationId": "remove_dependency_api_v1_tasks__task_id__dependencies_delete",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DependencyCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Remove Dependency"
      },
      "post": {
        "operationId": "add_dependency_api_v1_tasks__task_id__dependencies_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DependencyCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Add Dependency"
      }
    },
...
  }
}